requirements.txt
python-dateutil~=2.8.1
Flask-Login~=0.5.0
Flask~=1.1.2
gunicorn~=20.0.4
python-dotenv~=0.13.0
Flask-SQLAlchemy~=2.4.3
Werkzeug~=1.0.1
SQLAlchemy~=1.3.17